curriculo-rag-agent/pyproject.toml
[tool.poetry]
name = "curriculo-rag-agent"
version = "0.1.0"
description = "A beginner-friendly Agentic RAG system using FastAPI, Postgres (pgvector), and Gemini."
authors = ["Kevin <kev@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
sqlmodel = "^0.0.16"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.5"
google-generativeai = "^0.4.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
